ml-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
numpy==1.26.2
python-dotenv==1.0.0
pymongo==4.6.1
scipy==1.11.4
implicit==0.7.2
joblib==1.3.2
